refactor: snackbar usage
