+++ gradlock-front/app/(withLogout)/cadastro/aluno/page.tsx
@@ -1,70 +1,15 @@
 'use client';
 
-import {
-    Alert,
-    AlertProps,
-    Box,
-    Snackbar,
-    TextField,
-    Typography,
-} from '@mui/material';
+import { Box, TextField, Typography } from '@mui/material';
 import Logo from '@/components/ui/Logo';
 import AuthCard from '@/components/auth/AuthCard';
 import AuthButton from '@/components/auth/AuthButton';
 import { useState } from 'react';
-import { AuthService } from '@/api/services/authService';
 import { UserService } from '@/api/services/userService';
+import useSnackbar from '@/hooks/useSnackbar';
+import { CustomSnackbar } from '@/components/ui/CustomSnackBar';
 
 export default function RegisterStudent() {
-    type SnackbarState = {
-        open: boolean;
-        message: string;
-        severity: AlertProps['severity'];
-    };
-
-    const useSnackbar = () => {
-        const [snackbar, setSnackbar] = useState<SnackbarState>({
-            open: false,
-            message: '',
-            severity: 'success',
-        });
-
-        const showSnackbar = (
-            message: string,
-            severity: AlertProps['severity'] = 'success'
-        ) => {
-            setSnackbar({ open: true, message, severity });
-        };
-
-        const closeSnackbar = () => {
-            setSnackbar((prev) => ({ ...prev, open: false }));
-        };
-
-        return { snackbar, showSnackbar, closeSnackbar };
-    };
-
-    const CustomSnackbar = ({
-        snackbar,
-        onClose,
-    }: {
-        snackbar: SnackbarState;
-        onClose: () => void;
-    }) => (
-        <Snackbar
-            open={snackbar.open}
-            autoHideDuration={4000}
-            onClose={onClose}
-        >
-            <Alert
-                onClose={onClose}
-                severity={snackbar.severity}
-                variant="filled"
-            >
-                {snackbar.message}
-            </Alert>
-        </Snackbar>
-    );
-
     const { snackbar, showSnackbar, closeSnackbar } = useSnackbar();
     const [cpf, setCpf] = useState('');
     const [password, setPassword] = useState('');
@@ -105,7 +50,6 @@ export default function RegisterStudent() {
             showSnackbar(error?.message || 'Erro ao fazer cadastro!', 'error');
         }
     }
-
 
     return (
         <Box
