+++ gradlock-front/app/types/snackbar.ts
@@ -1,0 +1,7 @@
+import { AlertProps } from '@mui/material';
+
+export interface SnackbarState {
+    open: boolean;
+    message: string;
+    severity: AlertProps['severity'];
+}
